refactor: move `extractCurrentCodeWindow` from `index.ts` into `code-window.ts`
`extractCurrentCodeWindow` is a pure, side-effect-free helper that transforms file contents and a line coordinate into a windowed snippet. Hosting it in `index.ts` (the orchestration entry point) coupled callers to the top-level module and required exporting it for testability. Move it to a dedicated `src/code-window.ts` and colocate its unit tests in `src/code-window.test.ts`. `index.ts` now imports the helper instead of defining it.

Addresses manki-review review comment on PR #626
https://github.com/manki-review/manki/pull/626#discussion_r3146315046
